split server.ts
