+++ tslint-server/src/utils.ts
@@ -1,0 +1,16 @@
+import * as server from 'vscode-languageserver';
+
+export function trace(connection: server.IConnection, message: string, verbose?: string): void {
+	connection.tracer.log(message, verbose);
+}
+
+
+export function getErrorMessage(err: any, document: server.TextDocument): string {
+	let errorMessage = `unknown error`;
+	if (typeof err.message === 'string' || err.message instanceof String) {
+		errorMessage = <string>err.message;
+	}
+	let fsPath = server.Files.uriToFilePath(document.uri);
+	let message = `vscode-tslint: '${errorMessage}' while validating: ${fsPath} stacktrace: ${err.stack}`;
+	return message;
+}
